+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## 5th November 2020
+
+- Include default values
+- Add `OptionProvided`
+- Add `PropertyProvided`
+- Add typed `GetOption`
+- Include tests for the above
+
 ## 2nd November 2020
 
 - Initial commit